implement TrajectoryPlanner library
