+++ src/CMakeLists.txt
@@ -9,3 +9,4 @@ add_subdirectory(libStd_helper)
 add_subdirectory(libSimplified_model_controllers)
 add_subdirectory(libRobot_helper)
 add_subdirectory(libWhole_body_controllers)
+add_subdirectory(libTrajectory_planner)
